Implemented multiple statuses for songs and tracks in place of the older "published" flag [#3]
- Songs can be public, private, deleted or temporary. Status is accessed
via Song.statuses.public or Song.statuses[:public]. Temporary songs are
created in advance for the ajax controls on the MTrack page.
- Tracks can be public, private or deleted
- It is assumed that any user will evel delete Song/Track instances from
the database, and deletion merely sets the "deleted" status. Playlist
items and Mixes are deleted, instead. This is going to change.
- Added missing tracks and mixes to every song fixture, because of the
stricter validation checks
- Implemented deletion and destroy tests both for songs and for tracks

Unit tests run smoothly.