Reimplemented right the bare facebook connect integration code.
- Added a placeholder filter to check whether the /connect request is
really coming from a facebook server, because it's a huge security hole
to accept requests from users: facebook account wouldn't be affected,
but mewsic account would;

- Added error handling, both for request parameters and for user
validations;

- Added a migration that adds a facebook_uid string field to the users
table. String is used because as an external service, one should not trust
that IDS will EVER remain as big integers. It's remote data, and should
be treated as such;

- Added the `facebook_uid` field to the list of readonly ones to inhibit
mass-assignment, and added to the exception list of xss_terminate;

- Reworked top pane to show facebook avatar and username if connected,
or mewsic ones if not;

- Moved the javascript inclusion of the fb featureloader script into the
<head> of the document, and moved the javascript code into application.js;

- Added the ability to operate and test in environments without an
internet connection, even with a bit of lag (should be fixed): a console
message is shown in these cases;

- Added functional tests for the /sessions/connect action.